[turbofan] Print APIs are for debugging.
The Print APIs on the instruction model are for debugging. At debug
time, we cannot (easily) synthesize an output stream, hence the choice
of directing to stdout in those APIs.

The concern in https://codereview.chromium.org/2293413004/ is
addressed by the changes in pipeline.cc, using the various operator<<,
and does not require the changes in instruction.{h|cc}, and the
generalization of the Print APIs.

BUG=

Review-Url: https://codereview.chromium.org/2304423002
Cr-Commit-Position: refs/heads/master@{#39190}